Change atomics to use functions.
Use the atomic functions on int64's as opposed to
the atomic types.  This eliminates the vet noCopy warnings.

Also, remove the per config.Request mutex in favor of a
global mutex for the same reason.  No biggie here as contention
will always be for the same request as a result of thundering herd
resolutions.

Signed-off-by: Peter W. Morreale <pwmorreale@gmail.com>